+++ VaccinationScheduling/VaccinationScheduling.Shared/SchedulePrettier.cs
@@ -45,7 +45,7 @@ namespace VaccinationScheduling.Shared
                 foreach ((var schedule, int index) in jab2Hospital)
                 {
                     string job = $"Job {index + 1}";
-                    string line = ToLine(schedule.T2, global.TimeSecondDose);
+                    string line = ToLine((int)schedule.T2, global.TimeSecondDose);
                     stringBuilder.AppendLine($"{job.Pad()}{line}");
                 }
             }
